dice-protocols: lexicon: code refactoring
